Tried to make more comprehensive ... but not quite
Still can't handle '*' and there's still some char, or set of
chars, that are coming out as '\r'.  Heck, maybe that's the
char.  Ug.  Why the heck would somebody put a '\r' or '\n' in
a 'Subject' field?

And it can't handle doily UTF characters.

Also, tried to implement a mechanism to avoid collisions to
(possibly) existing files.  Not sure it's working, but I tested
it out so I think it works.

Changes to be committed:
	modified:   rename_emails.sh